+++ schema/claude/content_block.go
@@ -33,30 +33,30 @@ type CitationCharLocation struct {
 	CitedText string `json:"cited_text,omitempty"`
 
 	DocumentTitle string `json:"document_title,omitempty"`
-	DocumentIndex int64  `json:"document_index,omitempty"`
+	DocumentIndex int    `json:"document_index,omitempty"`
 
-	StartCharIndex int64 `json:"start_char_index,omitempty"`
-	EndCharIndex   int64 `json:"end_char_index,omitempty"`
+	StartCharIndex int `json:"start_char_index,omitempty"`
+	EndCharIndex   int `json:"end_char_index,omitempty"`
 }
 
 type CitationPageLocation struct {
 	CitedText string `json:"cited_text,omitempty"`
 
 	DocumentTitle string `json:"document_title,omitempty"`
-	DocumentIndex int64  `json:"document_index,omitempty"`
+	DocumentIndex int    `json:"document_index,omitempty"`
 
-	StartPageNumber int64 `json:"start_page_number,omitempty"`
-	EndPageNumber   int64 `json:"end_page_number,omitempty"`
+	StartPageNumber int `json:"start_page_number,omitempty"`
+	EndPageNumber   int `json:"end_page_number,omitempty"`
 }
 
 type CitationContentBlockLocation struct {
 	CitedText string `json:"cited_text,omitempty"`
 
 	DocumentTitle string `json:"document_title,omitempty"`
-	DocumentIndex int64  `json:"document_index,omitempty"`
+	DocumentIndex int    `json:"document_index,omitempty"`
 
-	StartBlockIndex int64 `json:"start_block_index,omitempty"`
-	EndBlockIndex   int64 `json:"end_block_index,omitempty"`
+	StartBlockIndex int `json:"start_block_index,omitempty"`
+	EndBlockIndex   int `json:"end_block_index,omitempty"`
 }
 
 type CitationWebSearchResultLocation struct {
